backend/requirements.txt
flask==2.3.3
flask-cors==4.0.0
sqlalchemy==2.0.23
pydantic==2.5.2
werkzeug==2.3.7
gunicorn==21.2.0
python-dotenv==1.0.0
